// src/app/models/cat.ts
export interface Cat {
    id: number;
    name: string;
    robe: string;
    dateOfBirth: string; // Utilisez une chaîne de caractères ou une date, en fonction de la manière dont vous souhaitez gérer les dates
    breed: string;
    urlProfil: string;
    urlProfilMother: string;
    urlProfilFather: string;
    sailliesExterieures: string;
    images: string[];
    eyeColor:string;
    sex: string;
    pedigree: string;
    profilId: number;
}